test: move mesh tests to test_meshing.py + deduplicate test
